move spanner code to it's package
